+++ pyo3_client_example_async/tests/conftest.py
@@ -15,7 +15,9 @@ def client():
 def test_server():
     # Start the server in a separate process
     server_path = "httpolars.test_utils.rate_limit_server"
-    process = subprocess.Popen(["uvicorn", f"{server_path}:app", "--host", "127.0.0.1", "--port", "8000"])
+    process = subprocess.Popen(
+        ["uvicorn", f"{server_path}:app", "--host", "127.0.0.1", "--port", "8000"]
+    )
     sleep(1)  # Give the server some time to start
 
     yield
